Moved more validation into base validator
