+++ src/SFA.DAS.ProviderApprenticeshipsService.Web.UnitTests/SFA.DAS.ProviderApprenticeshipsService.Web.UnitTests.csproj
@@ -85,20 +85,29 @@
     <Compile Include="Orchestrators\Commitments\WhenDeletingApprenticeship.cs" />
     <Compile Include="Orchestrators\Commitments\WhenGetCohorts.cs" />
     <Compile Include="Orchestrators\Commitments\WhenUploadingFile.cs" />
-    <Compile Include="Validation\WhenValidatingApprenticeshipBulkUploadViewModel.cs" />
+    <Compile Include="Validation\ApprenticeshipBulkUpload\WhenValidatingDateOfBirth.cs" />
+    <Compile Include="Validation\ApprenticeshipBulkUpload\WhenValidatingApprenticeshipBulkUploadViewModel.cs" />
+    <Compile Include="Validation\ApprenticeshipCreateOrEdit\ApprenticeshipValidadationTestBase.cs" />
+    <Compile Include="Validation\ApprenticeshipCreateOrEdit\WhenValidatingFirstName.cs" />
+    <Compile Include="Validation\ApprenticeshipCreateOrEdit\WhenValidatingDateOfBirth.cs" />
+    <Compile Include="Validation\ApprenticeshipCreateOrEdit\WhenValidatingTrainingCode.cs" />
+    <Compile Include="Validation\ApprenticeshipCreateOrEdit\WhenValidatingTrainingDates.cs" />
+    <Compile Include="Validation\ApprenticeshipCreateOrEdit\WhenValidatingCost.cs" />
+    <Compile Include="Validation\ApprenticeshipCreateOrEdit\WhenValidatingProviderReference.cs" />
+    <Compile Include="Validation\ApprenticeshipCreateOrEdit\WhenValidatingLastName.cs" />
     <Compile Include="WhenValidatingCsvRecord.cs" />
     <Compile Include="Orchestrators\BulkUpload\WhenCreatingRecordsBulkUpload.cs" />
     <Compile Include="Orchestrators\BulkUpload\WhenValidatingRecordsBulkUpload.cs" />
     <Compile Include="Orchestrators\BulkUpload\WhenValidatingBulkUploadFileAttributes.cs" />
     <Compile Include="Models\DateTimeViewModelTests.cs" />
-    <Compile Include="Validation\WhenValidatingApprenticeshipViewModelForApproval.cs" />
+    <Compile Include="Validation\ApprenticeshipApproval\WhenValidatingApprenticeshipViewModelForApproval.cs" />
     <Compile Include="Orchestrators\Commitments\WhenApproveCommitment.cs" />
     <Compile Include="Orchestrators\Commitments\WhenGettingCommitmentViewModel.cs" />
     <Compile Include="Orchestrators\Commitments\WhenFinishEditing.cs" />
     <Compile Include="StatusCalculator\WhenEditStatusIsWithEmployer.cs" />
     <Compile Include="StatusCalculator\WhenEditStatusIsWithBoth.cs" />
     <Compile Include="StatusCalculator\WhenEditStatusIsWithProvider.cs" />
-    <Compile Include="Validation\WhenValidatingApprenticeshipViewModel.cs" />
+    <Compile Include="Validation\ApprenticeshipCreateOrEdit\WhenValidatingUln.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
   <ItemGroup>
