+++ src/SFA.DAS.ProviderApprenticeshipsService.Web.UnitTests/Validation/ApprenticeshipBulkUpload/WhenValidatingApprenticeshipBulkUploadViewModel.cs
@@ -1,12 +1,14 @@
+using System;
 using FluentAssertions;
 
 using NUnit.Framework;
 
 using SFA.DAS.ProviderApprenticeshipsService.Web.Models;
+using SFA.DAS.ProviderApprenticeshipsService.Web.Models.Types;
 using SFA.DAS.ProviderApprenticeshipsService.Web.Validation;
 using SFA.DAS.ProviderApprenticeshipsService.Web.Validation.Text;
 
-namespace SFA.DAS.ProviderApprenticeshipsService.Web.UnitTests.Validation
+namespace SFA.DAS.ProviderApprenticeshipsService.Web.UnitTests.Validation.ApprenticeshipBulkUpload
 {
     [TestFixture]
     public class WhenValidatingApprenticeshipBulkUploadViewModel
@@ -17,7 +19,7 @@ namespace SFA.DAS.ProviderApprenticeshipsService.Web.UnitTests.Validation
         [SetUp]
         public void Setup()
         {
-            _validModel = new ApprenticeshipViewModel { ULN = "1001234567", FirstName = "TestFirstName", LastName = "TestLastName" };
+            _validModel = new ApprenticeshipViewModel { ULN = "1001234567", FirstName = "TestFirstName", LastName = "TestLastName", DateOfBirth = new DateTimeViewModel(DateTime.Now.AddYears(-16)) };
         }
 
         [Test]
@@ -42,8 +44,8 @@ namespace SFA.DAS.ProviderApprenticeshipsService.Web.UnitTests.Validation
             var result = _validator.Validate(_validModel);
             result.Errors.Count.Should().Be(2);
 
-            result.Errors[0].ErrorMessage.ShouldBeEquivalentTo("The <strong>Given names</strong> must be entered");
-            result.Errors[1].ErrorMessage.ShouldBeEquivalentTo("The <strong>Family name</strong> must be entered");
+            result.Errors[0].ErrorMessage.ShouldBeEquivalentTo("You must enter <strong>Given names</strong> that are no longer than 100 characters");
+            result.Errors[1].ErrorMessage.ShouldBeEquivalentTo("You must enter a <strong>Family name</strong> that's no longer than 100 characters");
         }
     }
 }
