+++ src/SFA.DAS.ProviderApprenticeshipsService.Web/Validation/Text/WebApprenticeshipValidationText.cs
@@ -1,4 +1,6 @@
-﻿namespace SFA.DAS.ProviderApprenticeshipsService.Web.Validation.Text
+﻿using System;
+
+namespace SFA.DAS.ProviderApprenticeshipsService.Web.Validation.Text
 {
     public class WebApprenticeshipValidationText : IApprenticeshipValidationErrorText
     {
@@ -51,11 +53,13 @@
             new ValidationMessage("The Learning start date is not valid", "LearnStartDate_01");
         public ValidationMessage LearnStartDate02 =>
             new ValidationMessage("The Learning start date must be entered and be in the format yyyy-mm-dd", "LearnStartDate_02");
+        public ValidationMessage LearnStartDate03 =>
+            new ValidationMessage("The start date must not be earlier than 1 May 2017", "LearnStartDate_03");
 
         public ValidationMessage LearnPlanEndDate01 =>
             new ValidationMessage("The Learning planned end date is not valid", "LearnPlanEndDate_01");
         public ValidationMessage LearnPlanEndDate02 =>
-            new ValidationMessage("The Learning planned end date must be entered and be in the format yyyy-mm-dd", "LearnPlanEndDate_02");
+            new ValidationMessage("The Learning planned end date is not valid", "LearnPlanEndDate_02");
         public ValidationMessage LearnPlanEndDate03 =>
             new ValidationMessage("The Learning planned end date must not be on or before the Learning start date", "LearnPlanEndDate_03");
         public ValidationMessage LearnPlanEndDate06 =>
@@ -66,59 +70,43 @@
         public ValidationMessage TrainingPrice02 =>
             new ValidationMessage("The cost must be 6 numbers or fewer, for example 25000", "TrainingPrice_02");
 
-        public ValidationMessage EPAOrgID01 =>
-            new ValidationMessage("The End point assessment id must not be more than 7 characters in length", "EPAOrgID_01");
-        public ValidationMessage EPAOrgID02 =>
-            new ValidationMessage("The End point assessment id must not be more than 7 characters in length", "EPAOrgID_02");
-
         public ValidationMessage ProviderRef01 =>
             new ValidationMessage("The Reference must be 20 characters or fewer", "ProviderRef_01");
 
-        // training type validation messages
+        public ValidationMessage EPAOrgID01 { get { throw new NotImplementedException(); } }
 
-        public ValidationMessage ProgType01 =>
-              new ValidationMessage("The Programme type must be entered and must not be more than 2 characters in length", "ProgType_01");
 
-        public ValidationMessage ProgType02 =>
-              new ValidationMessage("The Programme type is not a valid value", "ProgType_02");
+        public ValidationMessage EPAOrgID02 { get { throw new NotImplementedException(); } }
 
-        public ValidationMessage FworkCode01 =>
-              new ValidationMessage("The Framework code must not be more than 3 characters in length", "FworkCode_01");
+        public ValidationMessage FworkCode01 { get { throw new NotImplementedException(); } }
 
-        public ValidationMessage FworkCode02MustFworkCode =>
-              new ValidationMessage("The Framework code must be entered where the Programme Type is a framework", "FworkCode_02");
+        public ValidationMessage FworkCode02MustFworkCode { get { throw new NotImplementedException(); } }
 
-        public ValidationMessage FworkCode03MustNotFworkCode =>
-              new ValidationMessage("The Framework code must not be entered where the Programme Type is a standard", "FworkCode_03");
+        public ValidationMessage FworkCode03MustNotFworkCode { get { throw new NotImplementedException(); } }
 
-        public ValidationMessage FworkCode04 =>
-              new ValidationMessage("The Framework code is not a valid value", "FworkCode_04");
+        public ValidationMessage FworkCode04 { get { throw new NotImplementedException(); } }
 
-        public ValidationMessage PwayCode01 =>
-              new ValidationMessage("The Pathway code must not be more than 3 characters in length", "PwayCode_01");
+        public ValidationMessage ProgType01 { get { throw new NotImplementedException(); } }
 
-        public ValidationMessage PwayCode02 =>
-              new ValidationMessage("The Pathway code must be entered where the Programme Type is a framework", "PwayCode_02");
+        public ValidationMessage ProgType02 { get { throw new NotImplementedException(); } }
 
-        public ValidationMessage PwayCode03 =>
-              new ValidationMessage("The Pathway code must not be entered where the Programme Type is a standard", "PwayCode_03");
+        public ValidationMessage PwayCode01 { get { throw new NotImplementedException(); } }
 
-        public ValidationMessage PwayCode04 =>
-              new ValidationMessage("The Pathway code is not a valid value", "PwayCode_04");
+        public ValidationMessage PwayCode02 { get { throw new NotImplementedException(); } }
 
-        public ValidationMessage StdCode01 =>
-              new ValidationMessage("The Standard code must not be more than 5 characters in length", "StdCode_01");
+        public ValidationMessage PwayCode03 { get { throw new NotImplementedException(); } }
 
-        public ValidationMessage StdCode02 =>
-              new ValidationMessage("The Standard code must be entered where the Programme Type is a standard", "StdCode_02");
+        public ValidationMessage PwayCode04 { get { throw new NotImplementedException(); } }
 
-        public ValidationMessage StdCode03 =>
-              new ValidationMessage("The Standard code must not be entered where the Programme Type is a framework", "StdCode_03");
+        public ValidationMessage StdCode01 { get { throw new NotImplementedException(); } }
 
-        public ValidationMessage StdCode04 =>
-              new ValidationMessage("The Standard code is not a valid value", "StdCode_04");
+        public ValidationMessage StdCode02 { get { throw new NotImplementedException(); } }
 
-        public ValidationMessage TrainingCode01 =>
+        public ValidationMessage StdCode03 { get { throw new NotImplementedException(); } }
+
+        public ValidationMessage StdCode04 { get { throw new NotImplementedException(); } }
+
+        public ValidationMessage TrainingCode01 => 
             new ValidationMessage("Training code cannot be empty", "DefaultErrorCode");
     }
 }
